Refactor entrypoints
